requirements.txt
MetaTrader5
pandas
requests
python-dotenv
PyQt6
